requirements_ml.txt
# Core ML Libraries
numpy>=1.26.0
pandas>=2.0.3
scikit-learn>=1.3.0

# Advanced ML Models
xgboost==2.0.0
lightgbm==4.0.0
catboost==1.2

# Deep Learning
tensorflow==2.13.0
torch==2.0.1
keras==2.13.1

# Technical Analysis
ta-lib==0.4.26
pandas-ta==0.3.14b0
tulipy==0.4.0

# Trading & Market Data
alpaca-py==0.13.3
yfinance==0.2.28
ccxt==4.0.0

# Backtesting
vectorbt==0.25.3
backtrader==1.9.78.123
quantstats==0.0.59

# Optimization
scipy==1.11.1
cvxpy==1.3.2
optuna==3.3.0

# Feature Engineering
feature-engine==1.6.1
tsfresh==0.20.1

# Database
psycopg2-binary==2.9.7
redis==4.6.0
sqlalchemy==2.0.19

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
dash==2.11.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
joblib==1.3.2
tqdm==4.65.0

# Monitoring & Logging
prometheus-client==0.17.1
loguru==0.7.0

# API & Web
fastapi==0.101.0
uvicorn==0.23.2
websockets==11.0.3

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Jupyter for analysis
jupyter==1.0.0
jupyterlab==4.0.3
ipywidgets==8.1.0

# Time Series Specific
statsmodels==0.14.0
pmdarima==2.0.3
prophet==1.1.4

# Reinforcement Learning (Optional)
stable-baselines3==2.1.0
gym==0.26.2
tensorforce==0.6.5